App performance enhancements (#5022)
* fix spotlight cleanup

* remove modal navigation from router transactions

* cache gathered paths

* e2e updates

* update selection test

* lint, add comments

* event listener cleanup

* freeze tensorflow dep

* add < > shortcuts in help modal

* add test for shortcuts

* details

* clear frame stream lru cache

* fix shortcutToHelpItems

* add destroy items

* lint

* weak map resets and item deletions

* add max stream sizing configurations

* linting

* bugs

* clear stream cache on mouseleave

* better spotlight heuristics

* cleanup

* soft edit

* constants tweaks

* modal nav enhancements

* rm bytes sizing

* better aspect ratio thresholds

* tweaks

* config docs

* wait to stream for videos

* lru looker cache

* cleanup

* isReusable fixes

* cleanup

* move hook, fix e2e

* increase lru cache

* update screenshots

* rm python settings, add to options action

---------

+++ app/packages/spotlight/src/row.ts
@@ -119,8 +119,8 @@ export default class Row<K, V> {
     return this.#row[this.#row.length - ONE].item.id;
   }
 
-  destroy() {
-    this.#destroyItems();
+  destroy(destroyItems = false) {
+    destroyItems && this.#destroyItems();
     this.#aborter.abort();
   }
 
@@ -138,8 +138,11 @@ export default class Row<K, V> {
       throw new Error("row is not attached");
     }
 
+    if (!this.#config.retainItems) {
+      this.#destroyItems();
+    }
+
     this.#container.remove();
-    this.#destroyItems();
   }
 
   show(
